moved log_utilities to neuron
